# Faster concatenation of cubes with `AuxCoordFactory`s