--- a/docs/src/whatsnew/latest.rst
+++ b/docs/src/whatsnew/latest.rst
@@ -83,6 +83,9 @@ This document explains the changes made to Iris for this release
 #. `@bouweandela`_ updated :meth:`iris.cube.CubeList.concatenate` so it keeps
    ancillary variables and cell measures lazy. (:pull:`6010`)
 
+#. `@bouweandela`_ made :meth:`iris.cube.CubeList.concatenate` faster for cubes
+   that have coordinate factories. (:pull:`6038`)
+
 🔥 Deprecations
 ===============