@bouweandela 2024-07-03
Not enough to reach the threshold of a report, but this does make concatenate 10% faster:
| Change   | Before [d0801aaa]    | After [8635ff8b]    | Ratio   | Benchmark (Parameter)                                                                                  |
|----------|----------------------|---------------------|---------|--------------------------------------------------------------------------------------------------------|
|          | 132±0.8ms            | 118±1ms             | 0.90    | merge_concat.Concatenate.time_concatenate                                                              |